Emotion_app-main/requirements.txt
streamlit
groq
transformers
torch
sentencepiece
protobuf
gspread
google-auth
plotly
